backend/requirements.txt
fastapi
uvicorn[standard]
openai
tavily-python
python-dotenv
